GodisWatching/pyproject.toml
[project]
name = "godiswatching"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.68.0",
    "httpx>=0.28.1",
    "mss>=10.1.0",
    "ollama>=0.6.1",
    "pillow>=12.1.1",
    "plyer>=2.1.0",
    "python-dotenv>=1.2.2",
    "socksio>=1.0.0",
]
